South China Sea submap (南海小地图)

Build a small inset map of the South China Sea region, framed by a border box. Typically placed in the bottom-right corner of a map of China via geom_submap().

add_submap(
  data = NULL,
  colour = "black",
  linewidth = 0.3,
  fill = NA,
  border.colour = "black",
  border.linewidth = 0.5,
  bgcolor = "white",
  ...
)

Arguments

data

An sf object of the submap boundary. If NULL (default), the bundled bou_southsea dataset (the "十段线" national boundary in the South China Sea) is used.

colour, linewidth, fill

colour, line width and fill passed to ggplot2::geom_sf() for the boundary line.

border.colour, border.linewidth

colour and line width of the box drawn around the submap (panel.border).

bgcolor

background colour of the submap panel.

...

other parameters passed to ggplot2::geom_sf().

Value

A ggplot object.

See also