make_colorbar

make_colorbar

make_colorbar(
  at,
  labels = NULL,
  labeller = format,
  space = "right",
  width = 2,
  height = 1,
  col = NULL,
  alpha = 1,
  pretty = FALSE,
  equispaced = TRUE,
  tick.number = 7,
  tck = 0.3,
  tck.padding = 0,
  raster = FALSE,
  interpolate = FALSE,
  tri.upper = NA,
  tri.lower = NA,
  legend.line = element_line(size = 0.8),
  legend.box = element_rect(size = 0.5),
  hjust = 0.5,
  vjust = 0.5,
  size = 12,
  family = "Times",
  legend.text.location = c(0.5, 0.5),
  legend.margin = bb_margin(),
  title = NULL,
  legend.text = element_text(hjust = 0.5),
  legend.title = element_text(),
  padding.left = unit(2, "points"),
  padding.right = unit(2, "points"),
  ...,
  draw = FALSE,
  vp = NULL
)

colorkey_pos(space = "right")

Arguments

space

One of c('top', 'bottom', 'right', 'left')

legend.box

arrangement of multiple legends ("horizontal" or "vertical")

hjust, vjust

used in grid::grid.layout()

legend.margin

the margin around each legend (margin())

title

all title elements: plot, axes, legends (element_text(); inherits from text)

legend.text

cex:

col:

font:

fontfamily: The font family

fontface: The font face (bold, italic, ...)

lineheight:

legend.title

title of legend (element_text(); inherits from title)

padding.left, padding.right

padding in the left and right of the legend

...

additional element specifications not part of base ggplot2. In general, these should also be defined in the element tree argument.

draw

A logical value indicating whether graphics output should be produced.

vp

A Grid viewport object (or NULL).

Examples

library(rcolors)
library(ggplot2)

brks = c(-Inf, -1, 0, 1, 3, 6, 9, Inf)
brks = 1:10
nbrk = length(brks) - 1
cols = get_color(rcolors$amwg256, nbrk)

spaces = c("right", "left", "top", "bottom") #%>% set_names(., .)

g <- make_colorbar(
    at = brks, col = cols, height = 1,
    tck = 0.4,
    # padding.left = unit(2, "points"),
    # padding.right = unit(2, "points"),
    space = spaces[1],
    # legend.line = element_line(size = 0.1, linetype = 1, color = "black"),
    # legend.text = element_text(hjust = 0.5),
    legend.text.location = c(0.2, 0.5),
    # legend.text.just = c(0.5, 0.5),

    title = "Title",
    legend.title = element_text(hjust = 0, vjust = 1),
    # legend.box = element_rect(size = 0.5),
    # legend.line = element_line(size = 1),
    # legend.text = list(fontfamily = "Times", cex = 1.1),
    # hjust = 0.5
)
#> Error in make_colorbar(at = brks, col = cols, height = 1, tck = 0.4, space = spaces[1],     legend.text.location = c(0.2, 0.5), title = "Title", legend.title = element_text(hjust = 0,         vjust = 1), ): object 'cols' not found

# cowplot::plot_grid(plotlist = lst)

# Ipaper::write_fig({
#     grid.rect()
#     grid.draw(g)
# }, "a.pdf", 0.7, 4)

##
p <- ggplot(mtcars, aes(mpg, disp)) + geom_point()
p + g
#> Error in eval(expr, envir, enclos): object 'g' not found
# p2 = add_colorbar(p, g, width = 1)
# write_fig(p2, "a.pdf", 7, 4)