st_hatched_polygon

st_hatched_polygon

st_hatched_polygon(
  x,
  density = 10,
  angle = 45,
  fillOddEven = FALSE,
  type = "sf"
)

Arguments

x

SpatialPolygons* from library sp

density

the density of shading lines, in lines per inch. The default value of NULL means that no shading lines are drawn. A zero value of density means no shading nor filling whereas negative values and NA suppress shading (and so allow color filling).

angle

the slope of shading lines, given as an angle in degrees (counter-clockwise).

fillOddEven

logical controlling the polygon shading mode: see below for details. Default FALSE.

type

Controls the type of the returned object:

sf: sf object of MULTILINESTRING

sp: sp object of sp::SpatialLinesDataFrame()

Value

An hatched area for assigned type

References

https://github.com/statnmap/HatchedPolygons

https://statnmap.github.io/HatchedPolygons/

https://statnmap.github.io/HatchedPolygons/articles/leaflet_shading_polygon.html